Remove radarr type file
