+++ radarr/queue.go
@@ -1,0 +1,106 @@
+package radarr
+
+import (
+	"context"
+	"fmt"
+	"time"
+
+	"golift.io/starr"
+)
+
+// Queue is the /api/v3/queue endpoint.
+type Queue struct {
+	Page          int            `json:"page"`
+	PageSize      int            `json:"pageSize"`
+	SortKey       string         `json:"sortKey"`
+	SortDirection string         `json:"sortDirection"`
+	TotalRecords  int            `json:"totalRecords"`
+	Records       []*QueueRecord `json:"records"`
+}
+
+// QueueRecord is part of the activity Queue.
+type QueueRecord struct {
+	MovieID                 int64                  `json:"movieId"`
+	Languages               []*starr.Value         `json:"languages"`
+	Quality                 *starr.Quality         `json:"quality"`
+	CustomFormats           []interface{}          `json:"customFormats"` // probably []int64
+	Size                    float64                `json:"size"`
+	Title                   string                 `json:"title"`
+	Sizeleft                float64                `json:"sizeleft"`
+	Timeleft                string                 `json:"timeleft"`
+	EstimatedCompletionTime time.Time              `json:"estimatedCompletionTime"`
+	Status                  string                 `json:"status"`
+	TrackedDownloadStatus   string                 `json:"trackedDownloadStatus"`
+	TrackedDownloadState    string                 `json:"trackedDownloadState"`
+	StatusMessages          []*starr.StatusMessage `json:"statusMessages"`
+	DownloadID              string                 `json:"downloadId"`
+	Protocol                string                 `json:"protocol"`
+	DownloadClient          string                 `json:"downloadClient"`
+	Indexer                 string                 `json:"indexer"`
+	OutputPath              string                 `json:"outputPath"`
+	ID                      int64                  `json:"id"`
+	ErrorMessage            string                 `json:"errorMessage"`
+}
+
+// GetQueue returns a single page from the Radarr Queue (processing, but not yet imported).
+// WARNING: 12/30/2021 - this method changed. The second argument no longer
+// controls which page is returned, but instead adjusts the pagination size.
+// If you need control over the page, use radarr.GetQueuePage().
+// This function simply returns the number of queue records desired,
+// up to the number of records present in the application.
+// It grabs records in (paginated) batches of perPage, and concatenates
+// them into one list.  Passing zero for records will return all of them.
+func (r *Radarr) GetQueue(records, perPage int) (*Queue, error) {
+	return r.GetQueueContext(context.Background(), records, perPage)
+}
+
+// GetQueueContext returns a single page from the Radarr Queue (processing, but not yet imported).
+func (r *Radarr) GetQueueContext(ctx context.Context, records, perPage int) (*Queue, error) {
+	queue := &Queue{Records: []*QueueRecord{}}
+	perPage = starr.SetPerPage(records, perPage)
+
+	for page := 1; ; page++ {
+		curr, err := r.GetQueuePageContext(ctx, &starr.Req{PageSize: perPage, Page: page})
+		if err != nil {
+			return nil, err
+		}
+
+		queue.Records = append(queue.Records, curr.Records...)
+		if len(queue.Records) >= curr.TotalRecords ||
+			(len(queue.Records) >= records && records != 0) ||
+			len(curr.Records) == 0 {
+			queue.PageSize = curr.TotalRecords
+			queue.TotalRecords = curr.TotalRecords
+			queue.SortDirection = curr.SortDirection
+			queue.SortKey = curr.SortKey
+
+			break
+		}
+
+		perPage = starr.AdjustPerPage(records, curr.TotalRecords, len(queue.Records), perPage)
+	}
+
+	return queue, nil
+}
+
+// GetQueuePage returns a single page from the Radarr Queue.
+// The page size and number is configurable with the input request parameters.
+func (r *Radarr) GetQueuePage(params *starr.Req) (*Queue, error) {
+	return r.GetQueuePageContext(context.Background(), params)
+}
+
+// GetQueuePage returns a single page from the Radarr Queue.
+// The page size and number is configurable with the input request parameters.
+func (r *Radarr) GetQueuePageContext(ctx context.Context, params *starr.Req) (*Queue, error) {
+	var queue Queue
+
+	params.CheckSet("sortKey", "timeleft")
+	params.CheckSet("includeUnknownMovieItems", "true")
+
+	_, err := r.GetInto(ctx, "v3/queue", params.Params(), &queue)
+	if err != nil {
+		return nil, fmt.Errorf("api.Get(queue): %w", err)
+	}
+
+	return &queue, nil
+}
