+++ radarr/system.go
@@ -3,9 +3,36 @@ package radarr
 import (
 	"context"
 	"fmt"
+	"time"
 
 	"golift.io/starr"
 )
+
+// SystemStatus is the /api/v3/system/status endpoint.
+type SystemStatus struct {
+	Version           string    `json:"version"`
+	BuildTime         time.Time `json:"buildTime"`
+	StartupPath       string    `json:"startupPath"`
+	AppData           string    `json:"appData"`
+	OsName            string    `json:"osName"`
+	OsVersion         string    `json:"osVersion"`
+	Branch            string    `json:"branch"`
+	Authentication    string    `json:"authentication"`
+	SqliteVersion     string    `json:"sqliteVersion"`
+	URLBase           string    `json:"urlBase"`
+	RuntimeVersion    string    `json:"runtimeVersion"`
+	RuntimeName       string    `json:"runtimeName"`
+	MigrationVersion  int       `json:"migrationVersion"`
+	IsDebug           bool      `json:"isDebug"`
+	IsProduction      bool      `json:"isProduction"`
+	IsAdmin           bool      `json:"isAdmin"`
+	IsUserInteractive bool      `json:"isUserInteractive"`
+	IsNetCore         bool      `json:"isNetCore"`
+	IsMono            bool      `json:"isMono"`
+	IsLinux           bool      `json:"isLinux"`
+	IsOsx             bool      `json:"isOsx"`
+	IsWindows         bool      `json:"isWindows"`
+}
 
 // GetSystemStatus returns system status.
 func (r *Radarr) GetSystemStatus() (*SystemStatus, error) {
